Optimized input types.

+++ CFn.cpp
@@ -9,7 +9,7 @@ CFn::~CFn()
 {
 }
 
-CPoint CFn::GetPoint(double amplitude, double frequency, double m, double t, unsigned int Fm)
+CPoint CFn::GetPoint(unsigned int amplitude, float frequency, double m, float t, unsigned int Fm)
 {
 	double x = t;
 	double y = amplitude * sin(2 * 3.14 * (frequency + m * sin(2 * 3.14 * Fm * t) * t));
