+++ CPntAr.cpp
@@ -9,10 +9,10 @@ CPntAr::~CPntAr()
 {
 }
 
-unsigned int CPntAr::CalcPoints(double amplitude, double frequency, double m, unsigned int Fm)
+unsigned int CPntAr::CalcPoints(unsigned int amplitude, float frequency, double m, unsigned int Fm)
 {
 	Points.clear();
-	for (double t = 0; t <= 1000; t += 0.01)
+	for (float t = 0; t <= 1000; t += 0.01)
 	{
 		CPoint pt = GetPoint(amplitude, frequency, m, t, Fm);
 		Points.push_back(pt);
